add NoItemView with animation click link

+++ ToDoList/Views/ListView.swift
@@ -12,19 +12,28 @@ struct ListView: View {
   @EnvironmentObject var listViewModel: ListViewModel
   
   var body: some View {
-    List {
-      ForEach(listViewModel.items) { item in
-        ListRowView(item: item)
-          .onTapGesture {
-            withAnimation(.linear) {
-              listViewModel.updateItem(item: item)
-            }
+    
+    ZStack {
+      if listViewModel.items.isEmpty {
+        NoItemsView()
+          .transition(AnyTransition.opacity)
+          .animation(.easeInOut)
+      } else {
+        List {
+          ForEach(listViewModel.items) { item in
+            ListRowView(item: item)
+              .onTapGesture {
+                withAnimation(.linear) {
+                  listViewModel.updateItem(item: item)
+                }
+              }
           }
+          .onDelete(perform: listViewModel.deleteItem)
+          .onMove(perform: listViewModel.moveItem)
+        }
+        .listStyle(.plain)
       }
-      .onDelete(perform: listViewModel.deleteItem)
-      .onMove(perform: listViewModel.moveItem)
     }
-    .listStyle(.plain)
     .navigationTitle("📒 ToDo List")
     .navigationBarItems(
       leading: EditButton(),
@@ -35,7 +44,7 @@ struct ListView: View {
 }
 
 #Preview {
-  NavigationView {
+  NavigationStack {
     ListView()
   }
   .environmentObject(ListViewModel())
